fabric/param: Define socket drop rate as int
The socket drop rate is defined as a long, but an
int works just as well.  Convert it, so that we can
drop support for long parameters.

+++ prov/sockets/src/sock_util.h
@@ -43,7 +43,7 @@ extern const char sock_prov_name[];
 extern struct fi_provider sock_prov;
 extern int sock_pe_waittime;
 #if ENABLE_DEBUG
-extern long sock_dgram_drop_rate;
+extern int sock_dgram_drop_rate;
 #endif
 
 #define _SOCK_LOG_DBG(subsys, ...) FI_DBG(&sock_prov, subsys, __VA_ARGS__);
